Windows port: first executables link -- four programming examples build
Closes the last 15 unresolved externals:
- Simulator branches referencing umd::SimulationChip guarded off on
  Windows (get_platform_architecture.hpp, tt_cluster.cpp,
  core_descriptor.cpp): the simulation backend is CMake-excluded from
  Windows builds, so the branch now TT_THROWs ('TT_METAL_SIMULATOR is
  not supported on Windows builds') instead of pulling dead symbols
  into the link.
- Windows link stubs for the POSIX shared-memory socket transport
  appended to hd_socket_descriptor.cpp (always-built TU; the linker
  pulls it on demand): D2HSocket/H2DSocket referenced surface,
  NamedShm dtor, and the H2DSocket access shims all throw
  'not yet supported on Windows'. A real Windows shm port is a later
  milestone; no stub object is ever constructed on the reachable
  paths, so the defaulted dtors are safe.
- Explicit DeviceAddr->uint32_t narrowing in the loopback and
  eltwise_binary examples' braced runtime-arg initializers.

Built and present on disk (not executed):
  .build/windows-msvc/programming_examples/Release/
    metal_example_add_2_integers_in_compute.exe
    metal_example_loopback.exe
    metal_example_eltwise_binary.exe
    metal_example_matmul_single_core.exe

(cherry picked from commit 15c6a9cca3f79f17591d9d0155be366c4faab28b)

+++ tt_metal/programming_examples/eltwise_binary/eltwise_binary.cpp
@@ -141,8 +141,16 @@ int main(int /*argc*/, char** /*argv*/) {
 
         // Set the runtime arguments for the kernels. This also registers
         // the kernels with the program.
-        SetRuntimeArgs(program, reader, core, {src0_dram_buffer->address(), src1_dram_buffer->address(), n_tiles});
-        SetRuntimeArgs(program, writer, core, {dst_dram_buffer->address(), n_tiles});
+        // Buffer addresses are DeviceAddr (64-bit); runtime args are 32-bit words, so narrow
+        // explicitly (braced init rejects the implicit narrowing on MSVC).
+        SetRuntimeArgs(
+            program,
+            reader,
+            core,
+            {static_cast<uint32_t>(src0_dram_buffer->address()),
+             static_cast<uint32_t>(src1_dram_buffer->address()),
+             n_tiles});
+        SetRuntimeArgs(program, writer, core, {static_cast<uint32_t>(dst_dram_buffer->address()), n_tiles});
         SetRuntimeArgs(program, compute, core, {n_tiles});
 
         // We have setup the program. Now we queue the kernel for execution. The final argument is set to false. This indicates
